16177 bi_aggregation: Better visualization of aggregation errors
Aggregation errors are shown in the details of 'Aggr' services.
Previously, all aggregation errors were shown in the same tree
independant of whether they influence the state of the service or not,
which lead to confusion.
Now, two trees are shown, 'Aggregation problems affecting the state' and
'Aggregation problems not affecting the state'.

SUP-10277

Change-Id: I07d8e758a83bb1ed323beb3b9df0d5403d8ef7ba