16477 FIX Fix crash in filesystem related Nutanix services
This Werk fixes a crash in the Nutanix "NTNX Storage", "Storage Pool"
and "Container" services.

It also adds the option to update the naming of the "Prism Alerts",
"Prism Cluster", "Container" and "Storage Pool" services to be more consistent
with the other Nutanix services. The new names will be automatically used in new
installations.

Change-Id: I184f01fa99054aced1fe6028137d9c24c427ffd8
JIRA-Ref: CMK-16499