16630 FIX Prevent check_mail crash for "Move to subfolder" option
The active check check_mail can be configured to move mail messages from the inbox to a subfolder via the options "Forward mails as events to Event Console" > "Cleanup messages" > "Move to subfolder".
For IMAP the copying of mails crashed when there were no mails available in the inbox.
This is fixed to skipping the copy command in case there are no mails given.

SUP-17106

Change-Id: I0eae6f8390ec63576afe466ec64fef2f7b4f22d0