Make unit-tests independent of web/htdocs
In one of the following commits we will remove `./web/htdocs` and build
the files with `./packages/cmk-frontend`. This means we can no longer
assume that the files are available. The biggest problem is the theme
mechanic: The available themes are defined by folders in
`./web/htdocs/themes/`.

Our unit-tests should not be dependent on the themes available, so the
idea is to patch away most of this logic.

All the tests touched here do not test the theme logic. The code they
want to test indirectly depends on some theme loading logic.

Tests testing the theme logic are moved to integration-tests in
previous commits.

Change-Id: I70a1896715f2d98f3facec566c2e0fbc074b2b52