Use less invasive patching mechanism
Before: The code was copy-pasted in the repo and modified in place.
After: The code is downloaded and patched using standard mechanisms.

Going to newer versions of rules_foreign_cc should be easier with the
standard patching mechanism and having a patch file actually shows what
has been modified.

This has a further advantage of removing noise from our WORKSPACE.

CMK-16503

Change-Id: Ib84338e0e6621a9b4abf18c708e563b499804833