aws: Add filterable aws tags to agent output
This is a first step to allow a filtered import of all AWS tags
as either service or host labels. We chose not to add redundant
information where not already sent. That means that cloudwatch
metrics for example don't send the tags themselves but only
their "parent" summary section does.

In a second step, we will adapt the checks such that they parse
the sent tags as service labels (host labels are already parsed).

This commit will not change any customer-facing behavior except
for a slightly larger agent output that is not parsed.

CMK-16513

Change-Id: Ie414a59ab0939344c8e7bd1ed6d15d8e078e6778