16654 FIX Read-only internal folder name when editing folders
When disabling the "Hide internal folder names in Setup" setting, the internal
name must be set by the user when creating folders. Previously, the field was
still modifiable when editing the folder properties, however no changes to it
were saved.

This werk now marks the field as read only when editing the folder properties.

Change-Id: I53bbcc981082a4fc7b151eb0e38f746d28c0f01a
JIRA-Ref: CMK-16515