16626 trigger openapi-spec generation job during start, restart and reload
Werk 16501 introduced a command to start a background job which
triggers the regeneration of the API specification. This werk now
includes execution of this command also during omd start, restart,
and reload. With this mechanism the execution during `cmk-update-config`
is no longer needed.

Based on Werk 15724 the specification is now updated in these situations:

* Create the initial spec after a site has been created
* Update the spec after a site has been copied, restored or renamed
* Update the spec after the site has been updated
* Update the spec when the apache process is started, restarted or reloaded

Change-Id: Ic32c6afbfc08a43e03553de968dcd064cd685282