precompiled hosts file: make template accessible to linters
This makes the generated code accessable to linters and co,
while keeping the (performance) advantage of a compiled
file without IO to read a configuration.

Instead of having the whole file generated, we take a template
and only replace the definition of the CONFIG constant.

This uncovers quite a few problems.

We try to only make some small changes (drop unused imports)
to keep the change understandable.

Change-Id: If9c2b36b982515f95827ec1810b89f6b8707258a