Fix rate limit issue with Packer builds
We do an init everytime we build the cloud images. Because we use a
container the old init is not available anymore as it stored in
`~/.config/packer`. Now we place the config dir which contains the
plugins into the packer directory of the checkout. This should turn
`packer init .` into a no-op unless we clean the checkout directory.

Change-Id: Ic323ea11820cf6d27f76b1ea199e3e080d856d03