16612 notifications: Fix plugin permissions not loaded automatically
Previously, it was possible for users with the "Notification
configuration" permission to edit notification rules even if
they did not have the permission to the plugin that was being
used in the rule. When such a user edited such a rule, they
were able to overwrite the notification plugin with any plugin
they were allowed to use.

This werk stops users from editing rules that use notification
plugins they don't have access to.

CMK-16668

Change-Id: I73d45e650604fb62ff9469b1830be09d1772fc04