sites: fix validation error
Both "status_host" and "user_sync" do not need to be included in the config, so
they have been changed to OMITTED_FIELD.  In addition, if they are set and then
unset, they are then saved to file as None, so None has to also be a possible
value for both. Also, the secret field has been moved to the RemoteSiteModel.

Change-Id: I31192202998cdf505872f0f3dede9742a6f9f24f