make `dialog` errors verbose
We silence errors in dialog currently and just abort command. This can
be very frustrating to figure out what the error is. In the case of a
verbose flag, we print the error to the console.

Background: I'm running kitty as a terminal emulator. It sets the
variable `TERM=xterm-kitty` by default when I ssh into a machine. Now
machines I do not graphically log into do not have kitty installed. This
leads the `dialog` program to fail with the error message: "Error opening
terminal: xterm-kitty". Seeing that error would have been super helpful
to know what is going on.

Change-Id: Id603733741a2e2d1eb8f706c1db2468a7fd34283