16591 jenkins_version: New check for version of the queried Jenkins instance
A new check has been added which checks the version of the queried Jenkins instance.

The version is obtained through HTTP headers in the response of existing checks against Jenkins.
No additional configuration is required, besides the initial configuration of the Jenkins integration.

The version of the Jenkins instance will be compared against publically available release information.
In case an update is available the state will change to _warning_.

Change-Id: I6e241fb4e3ce76358fa2ab1ce4b8855b70b40a5c