16627 FIX kube_persistent_volume_claim: resolve KeyError crash when Volume parameters are configured
Before this update, configuring 'Volume parameters' in the PVC check plugin led to a crash. This
issue arose because the plugin was not designed to process trend size-related levels, which,
however, were anticipated by the general filesystem function handler shared among filesystem-related
check plugins. To resolve this, the general function has been modified to bypass trend computation
when a trend rule is not set, a scenario always applicable to the PVC check plugin.

Change-Id: Ief0e75d99c8883394e13441b074604f7c9f73118