16665 FIX Crash when accessing overridden built-in dashboard
Accessing a built-in dashboard after overriding it with a custom dashboard
could cause certain dashlets to crash.

For example, you could access the built-in dashboard by clicking the link
in Customize > Dashboards > Built-in. Another way to access the built-in
dashboard is for example by having a bookmark to it.

Now this crash no longer occurs and all dashlets render correctly.

CMK-16417

Change-Id: I964c3c36962a9ba787fec433af4c07be55ba3b62