15515 check_http: Soft deprecatation of old HTTP monitoring plug-in
The old plug-in is being deprecated in a soft way with this werk. Unlike
hard deprecation, the deprecated rule set "Check HTTP service" will remain
fully functional. However, new rules should only be created if absolutely
necessary, such as when experiencing issues with the new "Check HTTP web
service" implementation and needing to roll back to the old one.

Please note that the rule set will be hard deprecated in version 2.4.0,
meaning that you will no longer be able to create new rules. However, the
plug-in itself will remain available as this is a component of the
monitoring-plugins collection that comes with Checkmk.

Please know us know if you find any features that were present in the old
plug-in but are missing in the new one.

Change-Id: If6aa9feab5418965a3ecf46bcd49d5d2207e6214