15516 check_cert: New active check for advanced certificate monitoring
The _check_http_ plug-in was previously the only method to monitor
certificates out-of-the-box with Checkmk. With the new plug-in Checkmk
provides an extensive functionality to monitor certificates. This includes
but is not limited to certificates provided by the HTTP protocol.

With the new plug-in you can monitor all certificates provided through
a TCP connection to encrypt communication. This includes the monitoring
of

* validity times (max and remaining)
* issuer fields
* subject fields
* encryption algorithm
* alternative names
* response times
* public key algorithm and size
* serial number

As with the reworked plugin to monitor web services, you are able to
configure multiple services within a single rule.

Change-Id: I97c17721db25ac5b9e24c781920ed9deb2f3e2e0