tests.testlib.containers: Avoid disabling IPv6 for eth0
The function "execute_tests_in_container()" tried to disable
IPv6 support for the eth0 interface. We should not do that
and it might also cause the dockerized tests to fail with the
most recent version of docker if there is no eth0 interface
in the system.

Change-Id: I871beb5462df9450f15f1f6d4434be045a4e0965
(cherry picked from commit c6b265c521a599548f211c852253d0713160aef2)