Bazel output directories packaged as archives
Bazel 7 does not allow output directories
ERROR: dependency checking of directories is unsound

Bazel 7 does not allow naming files in foreign_cc
that are also included in a folder

Therefor we archive the output in tar archives and
remove the explicit naming of duplicated files.
The archives need to be unpacked for distribution.

This affects python3-modules, perl-modules, rrdtool and net-snmp.

Change-Id: I9672c015dd4a5bc0f27f35318871d1197d157d4c