Reenable disabled tests
Previously we had tests which queried the EC directly and ones which
interfaced with the core as a proxy. The direct ones did not work at
least since we executed the integration tests in the venv. There is
no permission to connect to the unix socket of the EC from the venv.

So this commit removes them and re-enables the tests going through the
core.

Change-Id: Ie7f5fcb0b7722e7780b96c7bb1ac878b859e22e8