PRO TIP: natively support password store in active check or special agent
The new SSC API allows you to pass the `Secret` "as is" or with its
`unsafe()` method.
The latter will pass the secret in plaintext, while the former will only
pass the ID of the password in the password store.

This change shows the required steps to go from the previous "magical"
hack obfuscating the passwords on the command line to the more straight
forward approach.

Note that we allow to pass the password explicitly as well, for
debugging purposes.

Change-Id: I90b240a16d7d1721ff1a1d3cf27d1695d9c7cff2