Avoid localizing the empty string (user-defined translations only)
The empty string is reserved for header data in PO files. Hence, the
localization of the empty string is the header data, which we certainly
do not want to display. Of course, static empty strings should not be
marked as localizable in the first place. However, user-supplied texts
are dynamic and can be empty (eg. descriptions of custom graphs).

CMK-16974

16355 FIX Custom graph objects: Avoid strange help texts in list views

Change-Id: I96a366cb9747d48ef23935f2a9298ad137d7565a