16633 FIX Logfile pattern analyzer: Fix crash for first rule without regex pattern
The "Logfile pattern analyzer" page crashed when the first "Logfile pattern" rule in place did not hold a regex pattern and a later rule did hold a regex pattern.
The rendering of the table of rules would crash with
```
Internal error: list index out of range
```

This is fixed and all rules are rendered as expected.

Change-Id: Ida61820e154dde3a72cdbd07f9e06ad0413324fe