tests/Makefile: Add target "container-debug" for test debugging
So far we had several "*-docker-debug" targets to run a container
for test debugging. Since all those targets use the same container,
this change adds a "container-debug" target for that purpose.

The help output was updated and does no longer mention the
"*-docker-debug" targets at all (but the targets are still kept
for backwards-compatibility).

Change-Id: Ideb318349c0334333b48c2861604fd9979a21402