Remove dynamic permission loading from REST API
It was added for unknown reasons while implementing the permission
system of the REST API. Currently it is not clear for what it is needed
while the call costs are high for each API call.

So we remove this now. It may have some side effects, which we have to
discover later on. In case we have to adjust this, we need to find a
more local solution to this to not slow down the API globally again.

CMK-16924

Change-Id: I4c192b77dae704ea7eb6228927ae6cf9d885ed51