Fallback for missing vite dev server
Previously it was possible that a customer could activate the "Inject
frontend_vue files via vite client" option in global settings and shut
themselves out of the User Interface.

We now introduced a "stage 1" script to load the scripts needed. If this
does not work it shows an error message and falls back to the original
javascript.

A pleasant side effect is, that now the "cmk-frontend-vue-ahr" magic
value is self contained in the cmk-frontend-vue project.

A downside is, that we have to mention the injection global option in
the error message, so this name is now leaking from cmk to the
package, and we have to share the "cmk-frontend-entrypoint-vue-stage1"
magic string to transport the fallback URL from the backend to the
frontend.

CMK-16468

Change-Id: I290e04ec7d96a8ce591cd3a1cc5a68194ce0ab2b