16756 host_config: introduce include_links option for Show all hosts endpoint
This werk introduces the flag 'include_links' for the 'Show all hosts'
host config endpoint. It toggles whether the links field should be populated
of the individual hosts. This serves as a mechanism to reduce the payload
generated. This field defaults to False and existing scripts therefore have
to be adjusted accordingly.

Change-Id: I5bf426c2e283f34a0b8cae2c58553a42bc2d9f0d