16659 ldap: allow manually updating locked status of users
When using the "Authentication Expiration" sync plug-in for LDAP, users can be
stuck in a locked state after too many failed login attempts.

With this werk it is now allowed to edit the "Disable password" option in the UI
(or "disable_login" in the REST API) for users managed by LDAP. Please note that
a sync with the LDAP will restore the original value.

Change-Id: I4ab0dad12279d004f49107661c80dd427fd00f0a
JIRA-Ref: SUP-18311