ipmi_sensors: Remove legacy perfometer
Note that this actually removes a perfometer in the UI in one special
case:
* Summary mode is activated in discovery parameters.
* There are sensors with a valid reading which have "amb" or "Ambient"
  in their name.

In this case, the ipmi_sensors plugin produces a metric named
"ambient_temp" (see `def _average_ambient_temperature`), which triggers
the legacy performeter removed by this commit. However, this metric is
not actually registered, so the perfometer has no units and it is very
unlikely that any customer actually knows what this perfometer displays
(since the plugin produces the metric only without any textual output).
Therefore, it should be ok to drop the perfometer in this particular
case.

Note that in non-summary mode, the ipmi_sensors plugin only produces
metrics for temperature sensors. These sensors are covered by the
standard temperature perfometer, so we don't change anything in this
case.

CMK-17098

Change-Id: I90eb0f07ed6ba3b1e8dd752fa1f40b926c2e0bee