omd: improve error message
If a distribution uses an init process other than the default, then
omd is unable to restart Apache. This specifically affects our
development containers, which use docker-init.

We now show that the restart is skipped. With a higher log level the
status command and it's exit code are shown.

CMK-16761

Change-Id: Icc5b7c09b415f9019ddd8fc01ee8c0672ff44d64