16781 FIX Bi: service/host_label group operator now has a default operator key "and"
When creating Bi rules via the REST-API, the host_label_groups and
the service_label_groups fields both accepted a list of conditions.
These conditions previously required an operator field.  However,
we realized that the operator field for the first item in these lists
really wasn't needed and had no affect. So that the user doesn't have
to provide this value, we have now added a default value for all
conditions in this list of "and".

CMK-16383

Change-Id: I6efb673590ddaf0c88fc9b5a9edf0f0bd2e9468f