Rules: service/host_label group operator now has a default operator key of 'and'
The operator field on the LabelGroupCondition was not set a required field, nor
did it have a load_default value.  This would cause an internal server error
if it was not provided.  Also, this Condition an item in a list and the first
item in this list does not require the operator key. To fix this, I've made
added a load_default value.

CMK-16383

Change-Id: Ieb37ecb4049fe402b301791f83d5152cbb7a0122