cmk-frontend: webpack: remove modules config
Previously we used a webpack config which allowed us to import files
from the `src/js/modules` folder as if they were globally installed
packages. This lead to name clashes and made it hard to understand what
files are actually imported.

We now stick to the current best practice: we either use relative
imports, or use a `@` alias for absolute imports.

CMK-16644

Change-Id: I4b32e9111389fc30a6f755998eea2af93e75bf97