16182 FIX bi: frozen node becomes CRIT and 'in service period' if the node disappeared
Previously, if the node of a frozen aggregation disappeared, it would
become UNKNOWN and `out of service period` instead of being removed from the tree.

Now, the node will become CRIT, but remain `in service period` in the same case.
The state was changed to alarm the user about the change in the aggregation tree better.
`out of service period` was removed because it does not apply to the entities that no longer exist.

CMK-15862

Change-Id: Id96b40cc7b8fcb3e1736c6d7f1230da46ab5fcca