16668 FIX Top list dashlet: Introduce best effort querying
In some rare cases, the top list dashlet does not display all available metrics
due to limitations in the metric system of Checkmk.

An example of this issue can be observed when using the `CPU utilization`
metric. When using Windows and Linux hosts at the same time, the services of
these hosts conflict with each other, potentially resulting in incomplete
results.

Now the top list dashlet will try to query all available metrics on a best
effort basis if the results are incomplete. If some metrics are still missing,
the dashlet displays an error message with the conflicting metrics.

This information can be used to identify the conflicting services and exclude
them by adding more filters to the dashlet.

CMK-17111

Change-Id: Ib000f19d514750522be6898a1e31d3b81f7ad828