Pagetypes: Pydantic + data classes: base classes
We implement a `config` property that converts the current spec
(`TypedDict`) into a data class. This allows for gradually migrating
all sub-classes.

This new property is only temporary. In a later commit, we will instead
pass a data class instance in the constructor. Also note that modifying
the object returned by `config` is pointless in the current state,
since changes are not persisted. However, this issue will solve itself
once the migration is complete.

CMK-17315

Change-Id: I6177ae564589f6d16c923d5901f4d7e6978acc0c