Licensing: Start trial period when switching to CCE or CME
When updating to an edition with mandatory licensing (CCE and CME), we want to give users enough time to apply the license when updating.

A CME updated to 2.3 will now start in 30-day trial period instead of directly entering the "free" license state if it otherwise would not be licensed.
However, in a distributed setup it is recommended to make use of the change in [werk #16193](https://checkmk.com/werk/16193) to apply the license before the update.

CMK-17297

Change-Id: I2af1c4f635fcb8a331e08e9b6b2243d95c3d2762