Remove UpdateActionState
Ideally, every update action needs be idempotent/stateless. However, sometimes
an action is only allowed to run once. In such a case, it is essential that a
component uses the version number of their data to ensure that the action is
applied correctly. If the schema for your data does not have a machine
readable version, please introduce one.

We might add some convenience functionality in the future make introducing a
version easier, but we are not there yet.

However, the versioning of components can't be part of `cmk/update/`. Imagine
you might need the version number of your data in some component. Will you
import that version from `cmk/update/`? -> No, your version goes with your
data into your component.



Change-Id: I676e5e6b6e38193b3cefc5139f3767b855cbc924