Pylint localization checkers: Take new API objects into account
The new rulesets and graphing & metrics API expose four types which mark
strings als localizable:
* Title
* Help
* Label
* Message

While this is of course user-friendly, it poses a challenge for linting,
since can now get collisions with other, non-related types with the same
name. For now, it is sufficient to restrict to callsites with one
argument only. Maybe we will find a better solution in the future.

CMK-17332

Change-Id: I1e2328ce0ac009a8f8b39249d2a9e849beb924c0