16669 FIX "This page was not found." error after operations on pagetypes
When performing certain operations, such as cloning, deleting, etc., on pagetypes
(views, dashboards, etc.) and when the "Find on this page ..." filter is active
on the page an error "This page was not found" was displayed on the page.

This issue has been resolved and the correct page is now displayed.

SUP-18463

Change-Id: I18557ce5764c61c7f29e5540c80f87609fb8f711