Unpin Pydantic and related packages
The Pydantic version we used before has an issue in the mypy plugin
which can lead to actual typing errors slipping through, see
https://github.com/pydantic/pydantic/pull/9008.

Note that the version we udpate to has this issue:
https://github.com/pydantic/pydantic/issues/9319
However, since a workaround is available, we still udpate to avoid
introducing new typing errors.

Also fix newly found issues. Apparently, polyfactory now has issues with
handling field aliases, so we switch to validation aliases, which are
sufficient in this case.

CMK-17006

Change-Id: I8b8ebb072f0c68b799d28c132c5f5605232db9c7