Remove old 'Event Console: Rewrite active config'
All customers must upgrade to 2.3.0, before upgrading to 2.4.0. Thus, any `UpdateAction` which was available in the 2.3.0 should no longer be needed in the 2.4.0.

These are some of the benefits to removing the `UpdateAction`s:

* Due to Werk 16408, all `UpdateAction`s should be reworked to do *most* of the heavy lifting in a pre-action. This way any failure can be rolled back.
* There are currently some anti-patterns in the update-checks, where some mistakes are made across all but a few plugins. See for instance:
Change-Id: I4f1b82404c9dd072d0adcd289379721b8fe0d07d
To handle these common mistakes, we need to introduce an API. The fewer `UpdateAction`s, the easier it is to introduce such an API.
* A dead action makes it harder for customers to troubleshoot their update, if something goes wrong.

CMK-17209

Change-Id: I0a8a1504aa3e548615d674bbfe7c8f1a00caccc4