Interface checks: More flexible handling of non-unicast packet rates
Previously, non-unicast packet rates were computed as the sum of broad-
and multicast packet rates. However, there are scenarios where the data
source only reports non-unicast data but not broad- and multicast data.

To account for such scenarios, we enable the interface checking code to
process non-unicast counters or rates directly as input. If this input
is not provided, the behavior is unchanged: Compute the non-unicast
rates as the sum of broad- and multicast if both are present.

SUP-18385

Change-Id: Iee5d4c0b29a627e23bb94bbd43c236a33e49fc65