16874 FIX host_config: skip host parent validation on host return
Prior to this werk, the REST API always validated if a specified
host parent is an existing host. This validation behavior differs
to the mechanism of the UI where it is possible that a host has
non existing hosts as parent. This werk modifies the behaviour of the
REST API to skip the host existence validation when a host is returned.

CMK-17301

Change-Id: I1ee28b3f8f09fbbc8f2968cca40f0233b3e36777