Untangle ip lookup error handler
We no longer add the failed lookups to the global variable,
but rather just collect them.

In the cases where we actually look at the global variable,
we now add the the warnings separately.

The logging that was previously a side effect of the collecting
is now done explicitly -- I am unsure as to when it is actually
desired, though.

Change-Id: Ibb9929ec75451641c4f02ae781f6c6b05e2e1f13