Refactoring: Simplify EOL handling in console.error
It doesn't make sense to go the extra mile to avoid emitting an EOL in
print(), just to add that EOL at every call site.

Note that one must be very careful when chaning such things in our code,
because the console code is effectively used for three totally different
purposes to emit:

   * fancy, perhaps colored output on the TTY
   * log lines in some log file
   * a fixed internal protocol

Sometimes one doesn't even know actual the purpose, because the code in
question is used in various scenarios.  This is quite suboptimal and should
really be disentangled, but this is something for future cleanups.

Change-Id: I681ee1b567d2f04c33c055f63f97f38ebada5252