piggyback: use the actual meta data for the source summary
The summarizer queried the piggyback raw data a second time
to create its summary. Since this operation is full of race
conditions, it is much better to transport the state observed
when the fetcher was executed.

We currently lack means to transport meta information
in the fetcher messages directly (other that an exception).
As a workaround we introduce a section into the payload.

Change-Id: Iaad18cce4415f03ed46b6de73dcc9c72e4497972