Fix KeyError for nested form_specs.DictGroup
This is only relevant for plugin developers.

In some cases when a `DictGroup` was used inside a FormSpec hierarchy an error would occur during validation of the relevant rule (e.g. when saving):
```
KeyError: 'DictGroup(title=<title>,help_text=<help_text>)'
```

Specifically this occurs when there are multiple `Transform`s applied outside the `DictGroup`, e.g. with multiple nested Dictionaries or when a `migrate`/`custom_validate` function is defined on the parent FormSPec of the `DictGroup`'s `Dictionary`.

CMK-17513

Change-Id: I2b5cc0840c25cdf0ad2ddb4027422fa5f2a9e81f