16766 FIX mknotifyd: use site names in service description
This update affects users monitoring the OMD Notify Connection service, regardless of whether the connection is encrypted.

Previously it was not possible to correctly monitor the Notification Spooler connection with TLS. For encrypted incoming connections, the IP address and port were unavailable because managed by stunnel using Unix sockets. Consequently, services for encrypted connections were named incorrectly and always displayed the IP address as 127.0.0.1.

To resolve this issue, we have updated the service naming. Now, the names of the involved sites will be shown in the format: OMD monitoring_site_name Notification Spooler connection to remote_site_name.

Impact:

- Currently monitored services (e.g., those configured with unencrypted connections) will continue to work as usual.
- Upon rediscovery, new services will be discovered and will adopt the new naming convention. For this reason, this werk is flagged as incompatible)
- To maintain service history, users can create a service description translation rule. This can be done in the configuration at: Setup > Agents > Access to agents > Translation of service descriptions. Use a regex or static string to map the old service names to the new ones.

Change-Id: If173105edd77f8b82ac1225c6721251af28c2e95