Split parametrized test into more descriptive individual tests
We should not have single-line tests that have a huge amount of
parameters as this is really hard to read and maintain.

Additionally we need to be careful when parametrizing tests with
dynamic data as this may unintentionally lead to dynamic test
names as well.

Change-Id: Idd365f8bfdcdbe4fea35f98fb87096c16e89cd16
(cherry picked from commit 6191f3353900acdbc87a8507c77a86b8a9df172d)