16590 Jenkins: Add metrics for memory, threads and active HTTP requests
The Jenkins check has been extended in order to be able to present
metrics about threads, memory and active HTTP requests.

The metric collection is done through the plugin called _Metrics_,
which has to be installed in Jenkins in order to make the data available.
Because of these additional requirements the metric collection has to
be manually enabled in the checkmk Jenkins integration configuration.
Please note that the proper retrieval of information requires the
monitoring user to have the _view_ permission for _Metrics_.

References CMK-16156

Change-Id: I7962f92a23aeb4069c1dc176305431ef45eba09f