16347 FIX Synthetic monitoring: Fix missing keyword metrics in test services
Before this werk, certain keyword runtime metrics were missing the output of the RMK test services.
This applied to both high-level and regex-matched keywords.
Specifically, this issue affected keywords nested in a TRY statement.
As a result of this werk, new keyword runtime metrics may appear in the test services.
Also, the names of existing metrics may change due to shifts in the underlying keyword IDs.

CMK-17640

Change-Id: Ie2d75f18f71b03f8a2fdb20eb285e82bce0f8521