17009 SEC XSS in inventory tree
Prior to this Werk an attacker with control over an agent was able to inject HTML in the output which was then rendered in the inventory tree of the coresponding host.

This problem exists only if the rule *Do hardware/software inventory* is set for the compromised agent/host.

We found this vulnerability internally.

**Affected Versions:**

* 2.3.0
* 2.2.0
* 2.1.0
* 2.0.0

**Mitigations**:
If you are unable to patch you can disable inventory scanning for all hosts.

**Indicators of Compromise:**
You can check `var/check_mk/inventory/` for inventories with embedded HTML.
This only indicates current 'attacks'.
Previous attacks (where the agent does not output the payload anymore) are not discoverable after some time (caching).

**Vulnerability Management:**
We have rated the issue with a CVSS Score of 6.5 (Medium) with the following CVSS vector:
`CVSS:3.1/AV:N/AC:L/PR:L/UI:R/S:C/C:L/I:L/A:L`
We assigned CVE-2024-5741 to this vulnerability.

**Changes:**
This Werk adds sanitation to the HTML output.

Change-Id: I5b93ac74128384c910fb17c54906bd62ee785d34