16786 FIX ldap_connection: empty failover_server list should be considered disabled
When configuring an LDAP connection via the REST-API, the default value
for the field failover_servers is an empty list. This value on the backend
should be considered disabled and just not set. Previously, we were saving
an empty list to the ldap config which was causing an error in the UI.

Change-Id: I3a0e62b6f7f367e588ad690c134ca85722239255