16839 mk-sql ensures instance name is correct
The plugin verifies whether the name of the instance at the
specified address matches the name indicated in the config.

If the name specified in the config does not match the
instance name that responds, the monitoring of this instance
will return an error.

Example: The config mistakenly indicates that the `SQLEXPRESS`
instance is located at `localhost:1433`, whereas the `MSSQLSERVER`
instance is actually located at this address.
Previously, the instance `MSSQLSERVER` were mistakenly monitored
as `SQLEXPRESS`.

Starting from this release, this error has been corrected.

You need to correct the config if your config contained bad
data.

Change-Id: I687f018ff753c259b624f966e8d3a7b6d97b44f5