16788 FIX notification_rules: encryption type now an enum
When creating a notification rule via the REST-API with the plugin
'asciimail', and configuring the 'enable_sync_smtp field', you could
previously set the subfield 'encryption' to any string.  Any string
is not a valid encryption type.

This werk addresses this issue by having this field be an enum of two
values which are

* ssl_tls
* starttls

This aligns with the configuration options found in the UI for
the same plugin.

Change-Id: I6f283e6b55d3365acd6f8011df04dbc8d5fd474b