17014 FIX notification_rules: auth value should be a required field when auth is enabled
Previously, when you created a notification rule via the REST-API using the
'mail' plugin, the value of the 'auth' field within the "enable_sync_smtp" schema
was not a required field. This is incorrect.  If 'auth' is enabled, then
the value of 'auth' should be a required field.  This werk now addresses
this issue.

e.g

```
"enable_sync_smtp": {
    "state": "enabled",
    "value": {
        "auth": {
            "state": "enabled",
            "value": {"method": "plaintext", "password": "1234", "user": "user_a"},
         },
         "encryption": "ssl_tls",
         "port": 25,
         "smarthosts": ["abc", "def"],
    },
}
```

Change-Id: I018650850e6d4f1f26e76adb2669ce442e9cfb11