Composition tests: Stop central site before tearing down remote
This hopefully avoids teardown failures such as

```
14:04:28  d6e2c8585c22: [31m[1m_______________ ERROR at teardown of test_automatic_host_removal _______________[0m
14:04:28  d6e2c8585c22:
14:04:28  d6e2c8585c22: [33m[31mFAILURE: [0mCrash report detected! ConnectionError: ('Connection aborted.', RemoteDisconnected('Remote end closed connection without response')).
14:04:28  d6e2c8585c22:                 See /omd/sites/comp_0_central/var/check_mk/crashes/gui/b4994570-2333-11ef-87eb-0242ac110009/crash.info for more details.
```

Note: an alternative approach would be to first remove the remote site
from the central site before tearing it down. However, we would first
have to remove all hosts and folders which are monitored on the remote
site. Otherwise, removing the remote site would fail. So simply stopping
the central site is much easier.

CMK-17722

Change-Id: Ifd47eb9715f9e02459d885943e3f5cd677e3c36b