16827 FIX page_handler errors appearing at the bottom of the page
Non-user errors that were thrown during the page_handler() call
were sometimes shown after a lot of content has been rendered,
depending on when the error was raised. Since they could appear
after having written a lot of html, we use a small javascript
helper function to move the error to the top of the page.

SUP-18178

Change-Id: Ic32b69912c2b0b6ecfe2a3d793ad99957a3c4613