testlib.containers: Run tests as testuser
The dockerized tests have been run as user root so far, which
caused some trouble like raising several MKGeneralException errors
when running the unit tests and is generally not recommended for
any integration and system tests.

This change fixes that by running the tests as testuser and making
the necessary adaptions to make this work.

Change-Id: I28be44ec305693e7c2e2e3e9a521d17f5c29348d