16876 FIX Nutanix agent: resolve verify error when environment REQUEST_CA_BUNDLE is set
Due to a specific behaviour in the requests Python library, the SSL verify disable
option does not work when the REQUEST_CA_BUNDLE environment variable is set. This
also affected the Nutanix agent. This werk resolves the error and therefore always
respects the no-cert-check flag. In addition, this werk also improves the error
message in case the agent fails when requesting the Nutanix API.

closes PR-703
CMK-17734

Change-Id: I04ddf2534f8104d144a99bbaf5940de5524f166f