Graph rendering: Respect lower vertical limit when rendering areas
Before, filled areas extended below the limit into the area of the time
axis label. Now, we cut off at the limit, similarly to the curve
rendering. When transitioning from above the limit to below or vice
versa, we render a pentagon instead of a quadrilateral. The lower edge
of the pentagon is exactly at the lower limit.

Note that we don't have to worry about the upper vertical range limit,
since the graph simply ends there. We can draw arbitrarily large
vertical values. They simply won't be rendered if they are above the
limit.

We will describe this change in a joint werk in a follow-up commit.

SUP-18828

Change-Id: Ib8007a413b6451b7e3b6e8c374b750437613edae