Graph rendering: Respect lower vertical limit when rendering curves
Before, curves were rendered below the limit into the area of the time
axis labels. Now, we cut off at the limit by connecting the last point
above the limit to a point directly on the limit. The time coordinate of
this point is computed from the slope and intersect of the line that
connects the last point above the limit with the first point below the
limit. And similarly for the opposite case where we transition from
below the limit to above the limit.

Note that we don't have to worry about the upper vertical range limit,
since the graph simply ends there. We can draw arbitrarily large
vertical values. They simply won't be rendered if they are above the
limit.

We will describe this change in a joint werk after also fixing the area
rendering, which suffers from the same bug.

SUP-18828

Change-Id: Iae8bb7b98cbf03164269a136fb2a255f98f0aabe